Move web app to web package & enable swagger api documentation
- TableInfo example
