+++ src/main/java/com/match/management/ServerApplication.java
@@ -1,4 +1,4 @@
-package com.match.management.server;
+package com.match.management;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
