+++ src/main/java/com/match/management/infrastructure/web/Match.java
@@ -1,0 +1,9 @@
+package com.match.management.infrastructure.web;
+
+import lombok.Data;
+
+@Data
+public class Match {
+    private String player1;
+    private String player2;
+}
