+++ src/main/java/com/match/management/infrastructure/web/TableState.java
@@ -1,0 +1,7 @@
+package com.match.management.infrastructure.web;
+
+public enum TableState {
+    UNASSIGNED,
+    ASSIGNED,
+    PLAYING
+}
